% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_gam.R
\name{anlz_gam}
\alias{anlz_gam}
\title{Fit a generalized additive model to a water quality time series}
\usage{
anlz_gam(moddat, mod = c("gam0", "gam1", "gam2", "gam6"), ...)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mod}{chr string indicating one of \code{gam0}, \code{gam1}, \code{gam2}, or \code{gam6}, see details,}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{\link[mgcv]{gam}} model object
}
\description{
Fit a generalized additive model to a water quality time series
}
\details{
One of four forms for the Generalized additive models (GAMs) can be selected to describes trends at the stations of interest.  The four models descibe the time components differently and represent increasing levels of complexity to describe the chlorophyll trend:

\describe{
 \item{gam0:}{chl ~ year + s(doy)}
 \item{gam1:}{chl ~ year + s(doy) + s(year)}
 \item{gam2:}{chl ~ year + s(doy) + s(year) + ti(doy, year)}
 \item{gam6:}{chl ~ year + s(doy) + s(year, k = large)}
}
For all models, \code{year} is measured as a continuous numeric variable for the annual effect (e.g., January 1st, 2000 is 2000.0, July 1st, 2000 is 2000.5, etc.) and \code{doy} is the day of year as a numeric value from 1 to 366.  All models include \code{year} as a linear effect.  The functions \code{\link[mgcv]{s}} model either \code{year} or \code{doy} as a smoothed, non-linear variable and \code{\link[mgcv]{ti}} models the interaction between the two.  The fourth model, \code{gam6}, is the same as \code{gam1} with the exception that the optimal amount of smoothing on \code{year} is not constrained by increasing the theoretical upper limit on the number of knots for \code{k}.  The upper limit of \code{k} was chosen as 12 times the number of years for the input data.
}
\examples{
library(dplyr)
tomod <- rawdat \%>\% 
  filter(station \%in\% 32) \%>\% 
  filter(param \%in\% 'chl')
anlz_gam(tomod, mod = 'gam2', trans = 'boxcox')
}
\seealso{
Other analyze: \code{\link{anlz_backtrans}},
  \code{\link{anlz_fit}}, \code{\link{anlz_prdday}},
  \code{\link{anlz_prd}}, \code{\link{anlz_smooth}},
  \code{\link{anlz_trans}}
}
\concept{analyze}
