% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_pred.R
\name{anlz_pred}
\alias{anlz_pred}
\title{Get predicted data from fitted GAMs across period of observation}
\usage{
anlz_pred(moddat = NULL, mods = NULL, length = 1000, ...)
}
\arguments{
\item{moddat}{input raw data, one station and parameter}

\item{mods}{optional list of model objects}

\item{length}{numeric indicating desired number of predictions}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{data.frame} with predictions
}
\description{
Get predicted data from fitted GAMs across period of observation
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_pred(tomod, trans = 'boxcox')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_pred(mods = mods)
}
\seealso{
Other analyze: \code{\link{anlz_backtrans}},
  \code{\link{anlz_fit}}, \code{\link{anlz_gam}},
  \code{\link{anlz_smooth}}, \code{\link{anlz_trans}}
}
\concept{analyze}
