% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_smooth.R
\name{anlz_smooth}
\alias{anlz_smooth}
\title{Return summary statistics for smoothers of all GAMs}
\usage{
anlz_smooth(rawdat = NULL, mods = NULL, ...)
}
\arguments{
\item{rawdat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{data.frame} with summary statistics for smoothers in each GAM
}
\description{
Return summary statistics for smoothers of all GAMs
}
\examples{
library(dplyr)

# fit models with function
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')

anlz_smooth(tomod, trans = 'boxcox')

# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_smooth(mods = mods)
}
\seealso{
Other analyze: \code{\link{anlz_gam}},
  \code{\link{anlz_trans}}
}
\concept{analyze}
