% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_trans.R
\name{anlz_trans}
\alias{anlz_trans}
\title{Transform response variable}
\usage{
anlz_trans(rawdat, trans = c("boxcox", "log10", "ident"))
}
\arguments{
\item{rawdat}{input raw data}

\item{trans}{chr string indicating desired type of transformation, one of \code{boxcox}, \code{log10}, or \code{ident} (no transformation)}
}
\value{
\code{rawdat} with the \code{value} column transformed as indicated
}
\description{
Transform response variable prior to fitting GAM
}
\examples{
library(dplyr)
tomod <- rawdat \%>\% 
  filter(station \%in\% 32) \%>\% 
  filter(param \%in\% 'chl')
anlz_trans(tomod, trans = 'boxcox')
}
\seealso{
Other analyze: \code{\link{anlz_gam}}
}
\concept{analyze}
