% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_trndseason.R
\name{anlz_trndseason}
\alias{anlz_trndseason}
\title{Estimate seasonal rates of change based on average estimates}
\usage{
anlz_trndseason(
  moddat = NULL,
  mods = NULL,
  doystr = 1,
  doyend = 364,
  justify = c("left", "right", "center"),
  win = 5,
  ...
)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{justify}{chr string indicating the justification for the trend window}

\item{win}{numeric indicating number of years to use for the trend window}

\item{...}{additional arguments passed to other methods}
}
\value{
A data frame of slope estimates and p-values for each year
}
\description{
Estimate seasonal rates of change based on average estimates
}
\details{
Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from \code{\link{anlz_mixmeta}}) for the seasonal averages (from \code{\link{anlz_avgseason}})
}
\examples{
library(dplyr)

# fit models with function
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_trndseason(tomod, trans = 'log10', doystr = 90, doyend = 180, justify = 'left', win = 5)
}
# use previously fitted list of models
trans <- 'log10'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_trndseason(mods = mods, trans = 'log10', doystr = 90, doyend = 180, justify = 'left', win = 5)
}
\seealso{
Other analyze: 
\code{\link{anlz_avgseason}()},
\code{\link{anlz_backtrans}()},
\code{\link{anlz_fit}()},
\code{\link{anlz_gam}()},
\code{\link{anlz_perchg}()},
\code{\link{anlz_prdday}()},
\code{\link{anlz_prd}()},
\code{\link{anlz_smooth}()},
\code{\link{anlz_trans}()}
}
\concept{analyze}
