% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_avgseason.R
\name{show_avgseason}
\alias{show_avgseason}
\title{Plot period (seasonal) averages from fitted GAM}
\usage{
show_avgseason(
  moddat = NULL,
  mods = NULL,
  doystr = 1,
  doyend = 364,
  yrstr = 2000,
  yrend = 2019,
  ylab,
  gami = c("gam0", "gam1", "gam2", "gam6"),
  ...
)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{yrstr}{numeric for starting year for trend model}

\item{yrend}{numeric for ending year for trend model}

\item{ylab}{chr string for y-axis label}

\item{gami}{Character indicating which GAM to plot, one of \code{gam0}, \code{gam1}, \code{gam2}, or \code{gam3}}

\item{...}{additional arguments passed to other methods}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot period (seasonal) averages from fitted GAM
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
show_avgseason(tomod, trans = 'boxcox', doystr = 90, doyend = 180, yrstr = 2000, yrend = 2019, 
     ylab = 'Chlorophyll-a (ug/L)', gami = 'gam2')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
show_avgseason(mods = mods, trans = 'boxcox', doystr = 90, doyend = 180, yrstr = 2000, yrend = 2019, 
     ylab = 'Chlorophyll-a (ug/L)', gami = 'gam2')
}
\seealso{
Other show: 
\code{\link{show_perchg}()},
\code{\link{show_prd3d}()},
\code{\link{show_prddoy}()},
\code{\link{show_prdseason}()},
\code{\link{show_prdseries}()},
\code{\link{show_trndseason}()}
}
\concept{show}
