% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_perchg.R
\name{show_perchg}
\alias{show_perchg}
\title{Plot percent change trends from GAM results for selected time periods}
\usage{
show_perchg(
  moddat = NULL,
  mods = NULL,
  baseyr,
  testyr,
  ylab,
  gami = c("gam0", "gam1", "gam2", "gam6"),
  ...
)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{baseyr}{numeric vector of starting years}

\item{testyr}{numeric vector of ending years}

\item{ylab}{chr string for y-axis label}

\item{gami}{Character indicating which GAM to plot, one of \code{gam0}, \code{gam1}, \code{gam2}, or \code{gam3}}

\item{...}{additional arguments passed to other methods}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot percent change trends from GAM results for selected time periods
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
show_perchg(tomod, gami = 'gam1', trans = 'boxcox', baseyr = 1990, testyr = 2016, 
     ylab = 'Chlorophyll-a (ug/L)')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
show_perchg(mods = mods, baseyr = 1990, testyr = 2016, ylab = 'Chlorophyll-a (ug/L)')
}
\seealso{
Other show: 
\code{\link{show_avgseason}()},
\code{\link{show_prd3d}()},
\code{\link{show_prddoy}()},
\code{\link{show_prdseason}()},
\code{\link{show_prdseries}()}
}
\concept{show}
