% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trndseason.R
\name{show_trndseason}
\alias{show_trndseason}
\title{Plot seasonal rates of change based on average estimates}
\usage{
show_trndseason(
  moddat = NULL,
  mods = NULL,
  doystr = 1,
  doyend = 364,
  justify = c("left", "right", "center"),
  win = 5,
  ylab,
  gami = c("gam0", "gam1", "gam2", "gam6"),
  ...
)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{justify}{chr string indicating the justification for the trend window}

\item{win}{numeric indicating number of years to use for the trend window}

\item{ylab}{chr string for y-axis label}

\item{gami}{Character indicating which GAM to plot, one of \code{gam0}, \code{gam1}, \code{gam2}, or \code{gam3}}

\item{...}{additional arguments passed to other methods}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot seasonal rates of change based on average estimates
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
show_trndseason(tomod, trans = 'log10', doystr = 90, doyend = 180, justify = 'left', win = 5,
     ylab = 'Slope chlorophyll-a (ug/L)', gami = 'gam2')
}
# use previously fitted list of models
trans <- 'log10'
mods <- list(
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
show_trndseason(mods = mods, trans = 'log10', doystr = 90, doyend = 180, justify = 'left', win = 5,
     ylab = 'Slope Chlorophyll-a (ug/L)', gami = 'gam2')
}
\seealso{
Other show: 
\code{\link{show_avgseason}()},
\code{\link{show_perchg}()},
\code{\link{show_prd3d}()},
\code{\link{show_prddoy}()},
\code{\link{show_prdseason}()},
\code{\link{show_prdseries}()}
}
\concept{show}
