% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscdat.R
\docType{data}
\name{sscdat}
\alias{sscdat}
\title{Daily-averaged suspended sediment concentration at Dumbarton Bridge}
\format{
A \code{data.frame} object with 7335 rows and 2 columns
\describe{
  \item{date}{Date}
  \item{ssc}{numeric, mg/L}
}
}
\usage{
sscdat
}
\description{
Daily-averaged suspended sediment concentration at Dumbarton Bridge
}
\details{
Data from \code{dat_prc} object in \url{https://github.com/fawda123/SFbaytrends}
}
\seealso{
Other utilities: 
\code{\link{rawdat}}
}
\concept{utilities}
\keyword{datasets}
