% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_avgseason.R
\name{anlz_avgseason}
\alias{anlz_avgseason}
\title{Extract period (seasonal) averages from fitted GAM}
\usage{
anlz_avgseason(moddat = NULL, mods = NULL, doystr = 1, doyend = 364, ...)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{...}{additional arguments passed to other methods}
}
\value{
A data frame of period averages
}
\description{
Extract period (seasonal) averages from fitted GAM
}
\examples{
library(dplyr)

# fit models with function
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_avgseason(tomod, trans = 'boxcox', doystr = 90, doyend = 180)
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_avgseason(mods = mods, doystr = 90, doyend = 180)
}
\seealso{
Other analyze: 
\code{\link{anlz_backtrans}()},
\code{\link{anlz_fit}()},
\code{\link{anlz_gam}()},
\code{\link{anlz_perchg}()},
\code{\link{anlz_prdday}()},
\code{\link{anlz_prd}()},
\code{\link{anlz_smooth}()},
\code{\link{anlz_trans}()},
\code{\link{anlz_trndseason}()}
}
\concept{analyze}
