% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_avgseason.R
\name{anlz_avgseason}
\alias{anlz_avgseason}
\title{Extract period (seasonal) averages from fitted GAM}
\usage{
anlz_avgseason(mod, doystr = 1, doyend = 364)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}
}
\value{
A data frame of period averages
}
\description{
Extract period (seasonal) averages from fitted GAM
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_avgseason(mod, doystr = 90, doyend = 180)
}
\seealso{
Other analyze: 
\code{\link{anlz_backtrans}()},
\code{\link{anlz_fit}()},
\code{\link{anlz_gam}()},
\code{\link{anlz_perchg}()},
\code{\link{anlz_prdday}()},
\code{\link{anlz_prd}()},
\code{\link{anlz_smooth}()},
\code{\link{anlz_trans}()},
\code{\link{anlz_trndseason}()}
}
\concept{analyze}
