% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_backtrans.R
\name{anlz_backtrans}
\alias{anlz_backtrans}
\title{Back-transform response variable}
\usage{
anlz_backtrans(dat)
}
\arguments{
\item{dat}{input data with \code{trans} argument}
}
\value{
\code{dat} with the \code{value} column back-transformed using info from the \code{trans} column
}
\description{
Back-transform response variable after fitting GAM
}
\details{
\code{dat} can be output from \code{\link{anlz_trans}} or \code{\link{anlz_pred}}
}
\examples{
library(dplyr)

tomod <- rawdat \%>\% 
  filter(station \%in\% 32) \%>\% 
  filter(param \%in\% 'chl')
dat <- anlz_trans(tomod, trans = 'log10')
anlz_backtrans(dat)

trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
dat <- anlz_pred(mods = mods)
anlz_backtrans(dat)
}
\seealso{
Other analyze: \code{\link{anlz_fit}},
  \code{\link{anlz_gam}}, \code{\link{anlz_predday}},
  \code{\link{anlz_pred}}, \code{\link{anlz_smooth}},
  \code{\link{anlz_trans}}
}
\concept{analyze}
