% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_fit.R
\name{anlz_fit}
\alias{anlz_fit}
\title{Return summary statistics for GAM fits}
\usage{
anlz_fit(moddat = NULL, mods = NULL, ...)
}
\arguments{
\item{moddat}{input raw data, one station and parameter}

\item{mods}{optional list of model objects}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{data.frame} with summary statistics for GAM fits
}
\description{
Return summary statistics for GAM fits
}
\examples{
library(dplyr)

# fit models with function
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_fit(tomod, trans = 'boxcox')
}

# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_fit(mods = mods)
}
\seealso{
Other analyze: \code{\link{anlz_gam}},
  \code{\link{anlz_pred}}, \code{\link{anlz_smooth}},
  \code{\link{anlz_trans}}
}
\concept{analyze}
