% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_mixmeta.R
\name{anlz_mixmeta}
\alias{anlz_mixmeta}
\title{Fit a mixed meta-analysis regresion model of trends}
\usage{
anlz_mixmeta(avgseason, yrstr, yrend)
}
\arguments{
\item{avgseason}{output from \code{\link{anlz_avgseason}}}

\item{yrstr}{numeric for starting year}

\item{yrend}{numeric for ending year}
}
\value{
A list of \code{\link[mixmeta]{mixmeta}} fitted model objects
}
\description{
Fit a mixed meta-analysis regresion model of trends
}
\examples{
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
avgseason <- anlz_avgseason(mods = mods, doystr = 90, doyend = 180)
anlz_mixmeta(avgseason)
}
