% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_perchg.R
\name{anlz_perchg}
\alias{anlz_perchg}
\title{Estimate percent change trends from GAM results for selected time periods}
\usage{
anlz_perchg(moddat = NULL, mods = NULL, baseyr, testyr, ...)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{baseyr}{numeric vector of starting years}

\item{testyr}{numeric vector of ending years}

\item{...}{additional arguments passed to other methods}
}
\value{
A data frame of summary results for change between the years.
}
\description{
Estimate percent change trends from GAM results for selected time periods
}
\details{
Working components of this function were taken from the \code{\link[baytrends]{gamDiff}} function in the baytrends package.
}
\examples{
library(dplyr)

# fit models with function
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_perchg(tomod, trans = 'boxcox', baseyr = 1990, testyr = 2016)
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_perchg(mods = mods, baseyr = 1990, testyr = 2016)
}
\seealso{
Other analyze: \code{\link{anlz_avgseason}},
  \code{\link{anlz_backtrans}}, \code{\link{anlz_fit}},
  \code{\link{anlz_gam}}, \code{\link{anlz_prdday}},
  \code{\link{anlz_prd}}, \code{\link{anlz_smooth}},
  \code{\link{anlz_trans}}
}
\concept{analyze}
