% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_prd.R
\name{anlz_prd}
\alias{anlz_prd}
\title{Get predicted data from fitted GAMs across period of observation}
\usage{
anlz_prd(mod)
}
\arguments{
\item{moddat}{input raw data, one station and parameter}

\item{mods}{optional list of model objects}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{data.frame} with predictions
}
\description{
Get predicted data from fitted GAMs across period of observation
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
  
mod <- anlz_gam(tomod, trans = 'log10')
anlz_prd(mod)
}
\seealso{
Other analyze: 
\code{\link{anlz_avgseason}()},
\code{\link{anlz_backtrans}()},
\code{\link{anlz_fit}()},
\code{\link{anlz_gam}()},
\code{\link{anlz_perchg}()},
\code{\link{anlz_prdday}()},
\code{\link{anlz_smooth}()},
\code{\link{anlz_trans}()},
\code{\link{anlz_trndseason}()}
}
\concept{analyze}
