% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_prd.R
\name{anlz_prd}
\alias{anlz_prd}
\title{Get predicted data from fitted GAMs across period of observation}
\usage{
anlz_prd(moddat = NULL, mods = NULL, ...)
}
\arguments{
\item{moddat}{input raw data, one station and parameter}

\item{mods}{optional list of model objects}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{data.frame} with predictions
}
\description{
Get predicted data from fitted GAMs across period of observation
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_prd(tomod, trans = 'boxcox')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_prd(mods = mods)
}
\seealso{
Other analyze: \code{\link{anlz_avgseason}},
  \code{\link{anlz_backtrans}}, \code{\link{anlz_fit}},
  \code{\link{anlz_gam}}, \code{\link{anlz_perchg}},
  \code{\link{anlz_prdday}}, \code{\link{anlz_smooth}},
  \code{\link{anlz_trans}}
}
\concept{analyze}
