% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_prdday.R
\name{anlz_prdday}
\alias{anlz_prdday}
\title{Get predicted data from fitted GAMs across period of observation, every day}
\usage{
anlz_prdday(mod)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}
}
\value{
a \code{data.frame} with predictions
}
\description{
Get predicted data from fitted GAMs across period of observation, every day
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_prdday(mod)
}
\seealso{
Other analyze: 
\code{\link{anlz_avgseason}()},
\code{\link{anlz_backtrans}()},
\code{\link{anlz_fit}()},
\code{\link{anlz_gam}()},
\code{\link{anlz_perchg}()},
\code{\link{anlz_prd}()},
\code{\link{anlz_smooth}()},
\code{\link{anlz_trans}()},
\code{\link{anlz_trndseason}()}
}
\concept{analyze}
