% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_smooth.R
\name{anlz_smooth}
\alias{anlz_smooth}
\title{Return summary statistics for smoothers of GAMs}
\usage{
anlz_smooth(moddat = NULL, mods = NULL, ...)
}
\arguments{
\item{moddat}{input raw data, one station and paramater}

\item{mods}{optional list of model objects}

\item{...}{additional arguments passed to other methods}
}
\value{
a \code{data.frame} with summary statistics for smoothers in each GAM
}
\description{
Return summary statistics for smoothers of GAMs
}
\details{
Results show the individual effects of the modelled components of each model as the estimated degrees of freedom (\code{edf}), the reference degrees of freedom (\code{Ref.df}), the test statistic (\code{F}), and significance of the component (\code{p-value}).  The significance of the component is in part based on the difference between \code{edf} and \code{Ref.df}.
}
\examples{
library(dplyr)

# fit models with function
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
anlz_smooth(tomod, trans = 'log10')
}
# use previously fitted list of models
trans <- 'log10'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_smooth(mods = mods)
}
\seealso{
Other analyze: 
\code{\link{anlz_avgseason}()},
\code{\link{anlz_backtrans}()},
\code{\link{anlz_fit}()},
\code{\link{anlz_gam}()},
\code{\link{anlz_perchg}()},
\code{\link{anlz_prdday}()},
\code{\link{anlz_prd}()},
\code{\link{anlz_trans}()},
\code{\link{anlz_trndseason}()}
}
\concept{analyze}
