% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_trndseason.R
\name{anlz_trndseason}
\alias{anlz_trndseason}
\title{Estimate seasonal rates of change based on average estimates}
\usage{
anlz_trndseason(
  mod,
  metfun = mean,
  doystr = 1,
  doyend = 364,
  justify = c("center", "left", "right"),
  win = 5,
  nsim = 10000,
  ...
)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{metfun}{function input for metric to calculate, e.g., \code{mean}, \code{var}, \code{max}, etc}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{justify}{chr string indicating the justification for the trend window}

\item{win}{numeric indicating number of years to use for the trend window}

\item{nsim}{numeric indicating number of random draws for simulating uncertainty}

\item{...}{additional arguments passed to \code{metfun}, e.g., \code{na.rm = TRUE)}}
}
\value{
A data frame of slope estimates and p-values for each year
}
\description{
Estimate seasonal rates of change based on average estimates
}
\details{
Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from \code{\link{anlz_mixmeta}}) for the seasonal metrics (from \code{\link{anlz_metseason}}).

Note that for left and right windows, the exact number of years in \code{win} is used. For example, a left-centered window for 1990 of ten years will include exactly ten years from 1990, 1991, ... , 1999.  The same applies to a right-centered window, e.g., for 1990 it would include 1981, 1982, ..., 1990 (if those years have data). However, for a centered window, picking an even number of years will always have an extra year to center the window exactly, e.g., a ten year window for 1990 will include eleven years from 1985 to 1995 so there is the same number of years to the left and right of center. A centered window with an odd number of years will always be centered and includes the exact number of years used in \code{win}.
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 34) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_trndseason(mod, doystr = 90, doyend = 180, justify = 'center', win = 8)
}
\seealso{
Other analyze: 
\code{\link{anlz_trans}()}
}
\concept{analyze}
