% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdat.R
\docType{data}
\name{rawdat}
\alias{rawdat}
\title{Raw data from San Francisco Estuary (South Bay)}
\format{
A \code{data.frame} object with 37301 rows and 8 columns
\describe{
  \item{date}{Date}
  \item{station}{int}
  \item{param}{chr}
  \item{value}{num}
  \item{doy}{num}
  \item{dec_time}{num}
  \item{yr}{num}
  \item{mo}{Ord.factor}
}
}
\usage{
rawdat
}
\description{
Raw data from San Francisco Estuary (South Bay)
}
\details{
Data from \code{datprc} object in \url{https://github.com/fawda123/SFbaytrends}
}
\concept{utilities}
\keyword{datasets}
