% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_avgseason.R
\name{show_avgseason}
\alias{show_avgseason}
\title{Plot period (seasonal) averages from fitted GAM}
\usage{
show_avgseason(mod, doystr = 1, doyend = 364, yrstr = 2000, yrend = 2019, ylab)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{yrstr}{numeric for starting year for trend model}

\item{yrend}{numeric for ending year for trend model}

\item{ylab}{chr string for y-axis label}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot period (seasonal) averages from fitted GAM
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 34) \%>\%
  filter(param \%in\% 'chl')
  
mod <- anlz_gam(tomod, trans = 'ident')

show_avgseason(mod, doystr = 90, doyend = 180, yrstr = 2000, yrend = 2019, 
     ylab = 'Chlorophyll-a (ug/L)')
}
\concept{show}
