% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_prddoy.R
\name{show_prddoy}
\alias{show_prddoy}
\title{Plot predictions for GAMs against day of year}
\usage{
show_prddoy(predin, ylab, nfac = NULL)
}
\arguments{
\item{predin}{data frame input to plot, output from \code{\link{anlz_pred}}}

\item{ylab}{chr string for y-axis label}

\item{nfac}{numeric indicating column number for facets, passed to \code{ncol} argument of \code{\link[ggplot2]{facet_wrap}}, defaults to number of model outputs in \code{predin}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot predictions for GAMs against day of year
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
prds <- anlz_pred(tomod, trans = 'boxcox')
show_prddoy(prds, ylab = 'Chlorophyll-a (ug/L)')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans),
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
)

prds <- anlz_pred(mods = mods)

show_prddoy(prds, ylab = 'Chlorophyll-a (ug/L)')
}
\concept{show}
