% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_prddoy.R
\name{show_prddoy}
\alias{show_prddoy}
\title{Plot predictions for GAMs against day of year}
\usage{
show_prddoy(moddat = NULL, mods = NULL, ylab, nfac = NULL, ...)
}
\arguments{
\item{moddat}{input raw data, one station and parameter}

\item{mods}{optional list of model objects}

\item{ylab}{chr string for y-axis label}

\item{nfac}{numeric indicating column number for facets, passed to \code{ncol} argument of \code{\link[ggplot2]{facet_wrap}}, defaults to number of model outputs in \code{predin}}

\item{...}{additional arguments passed to other methods}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot predictions for GAMs against day of year
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
show_prddoy(moddat = tomod, ylab = 'Chlorophyll-a (ug/L)', trans = 'boxcox')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans),
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
)

show_prddoy(mods = mods, ylab = 'Chlorophyll-a (ug/L)')
}
\seealso{
Other show: \code{\link{show_prd3d}},
  \code{\link{show_prdseason}},
  \code{\link{show_prdseries}}
}
\concept{show}
