% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_prdseason.R
\name{show_prdseason}
\alias{show_prdseason}
\title{Plot predictions for GAMs over time, by season}
\usage{
show_prdseason(moddat = NULL, mods = NULL, ylab, nfac = NULL, ...)
}
\arguments{
\item{moddat}{input raw data, one station and parameter}

\item{mods}{optional list of model objects}

\item{ylab}{chr string for y-axis label}

\item{nfac}{numeric indicating column number for facets, passed to \code{ncol} argument of \code{\link[ggplot2]{facet_wrap}}, defaults to one}

\item{...}{additional arguments passed to other methods}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot predictions for GAMs over time, by season
}
\examples{
library(dplyr)

# get predictions for all four gams
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')
\dontrun{
show_prdseason(moddat = tomod, ylab = 'Chlorophyll-a (ug/L)', trans = 'boxcox')
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans),
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
)

show_prdseason(moddat = tomod, mods = mods, ylab = 'Chlorophyll-a (ug/L)')
}
\seealso{
Other show: \code{\link{show_perchg}},
  \code{\link{show_prd3d}}, \code{\link{show_prddoy}},
  \code{\link{show_prdseries}}
}
\concept{show}
