% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_prdseason.R
\name{show_prdseason}
\alias{show_prdseason}
\title{Plot predictions for GAMs over time, by season}
\usage{
show_prdseason(mod, ylab)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{ylab}{chr string for y-axis label}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot predictions for GAMs over time, by season
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 34) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
show_prdseason(mod, ylab = 'Chlorophyll-a (ug/L)')
}
\seealso{
Other show: 
\code{\link{show_avgseason}()},
\code{\link{show_perchg}()},
\code{\link{show_prd3d}()},
\code{\link{show_prddoy}()},
\code{\link{show_prdseries}()},
\code{\link{show_trndseason}()}
}
\concept{show}
