% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_prdseries.R
\name{show_prdseries}
\alias{show_prdseries}
\title{Plot predictions for GAMs over time series}
\usage{
show_prdseries(mod, ylab, alpha = 0.7)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{ylab}{chr string for y-axis label}

\item{alpha}{numeric from 0 to 1 indicating line transparency}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot predictions for GAMs over time series
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 32) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')

show_prdseries(mod, ylab = 'Chlorophyll-a (ug/L)')
}
\seealso{
Other show: 
\code{\link{show_avgseason}()},
\code{\link{show_perchg}()},
\code{\link{show_prd3d}()},
\code{\link{show_prddoy}()},
\code{\link{show_prdseason}()},
\code{\link{show_trndseason}()}
}
\concept{show}
