% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sumtrndseason.R
\name{show_sumtrndseason}
\alias{show_sumtrndseason}
\title{Plot seasonal rates of change based on average estimates for multiple window widths}
\usage{
show_sumtrndseason(
  mod,
  doystr = 1,
  doyend = 364,
  justify = c("center", "left", "right"),
  win = 5:15,
  txtsz = 6,
  cols = c("lightblue", "lightgreen")
)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{justify}{chr string indicating the justification for the trend window}

\item{win}{numeric vector indicating number of years to use for the trend window}

\item{txtsz}{numeric for size of text labels inside the plot}

\item{cols}{vector of low/high colors for trends}
}
\value{
A data frame of slope estimates and p-values for each year
}
\description{
Plot seasonal rates of change based on average estimates for multiple window widths
}
\details{
This function plots output from \code{\link{anlz_sumtrndseason}}.
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 34) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
show_sumtrndseason(mod, doystr = 90, doyend = 180, justify = 'center', win = 5:15)
}
\concept{show}
