% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trndseason.R
\name{show_trndseason}
\alias{show_trndseason}
\title{Plot seasonal rates of change based on average estimates}
\usage{
show_trndseason(
  mod,
  doystr = 1,
  doyend = 364,
  type = c("log", "approx"),
  justify = c("left", "right", "center"),
  win = 5,
  ylab
)
}
\arguments{
\item{mod}{input model object as returned by \code{\link{anlz_gam}}}

\item{doystr}{numeric indicating start Julian day for extracting averages}

\item{doyend}{numeric indicating ending Julian day for extracting averages}

\item{justify}{chr string indicating the justification for the trend window}

\item{win}{numeric indicating number of years to use for the trend window}

\item{ylab}{chr string for y-axis label}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot seasonal rates of change based on average estimates
}
\examples{
library(dplyr)

# data to model
tomod <- rawdat \%>\%
  filter(station \%in\% 34) \%>\%
  filter(param \%in\% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
show_trndseason(mod, doystr = 90, doyend = 180, justify = 'left', win = 5,
     ylab = 'Slope Chlorophyll-a (ug/L/yr)')
}
\seealso{
Other show: 
\code{\link{show_avgseason}()},
\code{\link{show_perchg}()},
\code{\link{show_prd3d}()},
\code{\link{show_prddoy}()},
\code{\link{show_prdseason}()},
\code{\link{show_prdseries}()}
}
\concept{show}
